22:55

+++ Models/Category.php
@@ -8,7 +8,6 @@ class Category
 
 	private $con;
 		//Atributos del objeto usuario
-        public $id_categoria;
         public $nom_categoria;
         public $descripcion;
 	//Método de conexión a SGBD.
@@ -53,7 +52,7 @@ class Category
 	}
 	//Este método selecciona todas las tuplas de la tabla
 	//usuario en caso de error se muestra por pantalla.
-	public function get()
+	public function get($option)
 	{
 		try
 		{
@@ -61,25 +60,35 @@ class Category
 			FROM v_mostrarCategoria";
 			//Sentencia SQL para selección de datos.
 			$stm = sqlsrv_query($this->con, $sql);
+			if($option==1){
+				while($r = sqlsrv_fetch_array($stm)){
+					$id_categoria = $r['id_categoria'];
+					$nom_categoria = $r['nom_categoria'];
+					$descripcion = $r['descripcion'];
+					?>
+						<tr class="bg-light">
+							<td><?php echo $id_categoria; ?></td>
+							<td><?php echo $nom_categoria; ?></td>
+							<td><?php echo $descripcion?></td>
+							<td>
+								<button type="button" class="btn btn-success editbtn" data-toggle="modal" data-target="#editar">Editar</button>
+								<button type="button" class="btn btn-danger mt-0">
+									<a class="link" onclick="javascript:return confirm('¿Seguro de eliminar este registro?');" href="?c=categoryController&a=delete&id=<?php echo $r['id_categoria']; ?>">Eliminar</a>
+								</button>
+							</td>
+						</tr>
+					<?php 
+				}  				
 
-			while($r = sqlsrv_fetch_array($stm)){
-				$id_categoria = $r['id_categoria'];
-				$nom_categoria = $r['nom_categoria'];
-				$descripcion = $r['descripcion'];
-				?>
-					<tr class="bg-light">
-						<td><?php echo $id_categoria; ?></td>
-						<td><?php echo $nom_categoria; ?></td>
-						<td><?php echo $descripcion?></td>
-						<td>
-							<button type="button" class="btn btn-success editbtn" data-toggle="modal" data-target="#editar">Editar</button>
-							<button type="button" class="btn btn-danger mt-0">
-								<a class="link" onclick="javascript:return confirm('¿Seguro de eliminar este registro?');" href="?c=categoryController&a=delete&id=<?php echo $r['id_categoria']; ?>">Eliminar</a>
-							</button>
-						</td>
-					</tr>
-				<?php 
-		    }  
+			}else if($option==2){
+				while($r = sqlsrv_fetch_array($stm)){
+					$id_categoria = $r['id_categoria'];
+					$nom_categoria = $r['nom_categoria'];
+					?>
+						<option value="<?php echo $id_categoria = $r['id_categoria']; ?>"><?php echo $nom_categoria = $r['nom_categoria'];; ?></option>
+					<?php 
+				}
+			}
 	    }
 		catch(Exception $e)
 		{
@@ -87,22 +96,7 @@ class Category
 			die($e->getMessage());
 		}
 	}
-	public function select(){
-		$sql = "select id_categoria, nom_categoria, descripcion
-		FROM v_mostrarCategoria";
-		//Sentencia SQL para selección de datos.
-		$stm = sqlsrv_query($this->con, $sql);
-		while($r = sqlsrv_fetch_array($stm)){
-			$id_categoria = $r['id_categoria'];
-			$nom_categoria = $r['nom_categoria'];
-			$descripcion = $r['descripcion'];
-			?>
-<option value="<?php echo $id_categoria = $r['id_categoria']; ?>"><?php echo $nom_categoria = $r['nom_categoria'];; ?></option>
-			<?php 
-		}
-	//	return $r;
-		
-	}
+
 	public function delete($id_categoria)
 	{
 		try
